+++ attributes/user_install.rb
@@ -1,6 +1,6 @@
-default['rvm']['keyserver'] = 'hkp://keys.gnupg.net'
-default['rvm']['recv-keys'] = '409B6B1796C275462A1703113804BB82D39DC0E3'
+default['rvm_sl']['keyserver'] = 'hkp://keys.gnupg.net'
+default['rvm_sl']['recv-keys'] = '409B6B1796C275462A1703113804BB82D39DC0E3'
 
-default['rvm']['user']['name'] = 'admin'
-default['rvm']['user']['password'] = 'admin'
-default['rvm']['user']['dir'] = '/home/admin'
+default['rvm_sl']['user']['name'] = 'admin'
+default['rvm_sl']['user']['password'] = 'admin'
+default['rvm_sl']['user']['dir'] = '/home/admin'