+++ metadata.rb
@@ -1,7 +1,7 @@
-name             'rvm'
+name             'rvm_sl'
 maintainer       'David Saenz Tagarro'
 maintainer_email 'david.saenz.tagarro@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures rvm'
+description      'Installs/Configures rvm_sl'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'