+++ recipes/rubies.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: rvm
+# Cookbook Name:: rvm_sl
 # Recipe:: rubies
 #
 # Copyright 2015, David Saenz Tagarro
@@ -7,12 +7,12 @@
 # All rights reserved - Do Not Redistribute
 #
 
-user_name = node['rvm']['user']['name']
+user_name = node['rvm_sl']['user']['name']
 home = "/home/#{user_name}"
 
-ruby_version = node['rvm']['rubies']['version']
+ruby_version = node['rvm_sl']['rubies']['version']
 
-group 'rvm' do
+group 'rvm_sl' do
   members [user_name, 'root']
   append true
 end
@@ -25,12 +25,12 @@ end
 execute 'install_ruby' do
   environment 'HOME' => '/home/vagrant'
   command <<-END
-    bash -l -c 'rvm install #{ruby_version}'
-    bash -l -c 'rvm --default use #{ruby_version}'
+    bash -l -c 'rvm_sl install #{ruby_version}'
+    bash -l -c 'rvm_sl --default use #{ruby_version}'
   END
 end
 
-execute 'chown_rvm_dir' do
-  command "chown -R #{user_name}:rvm #{home}/.rvm"
+execute 'chown_rvm_sl_dir' do
+  command "chown -R #{user_name}:rvm_sl #{home}/.rvm_sl"
   action :run
 end