+++ spec/rubies_spec.rb
@@ -1,12 +1,12 @@
 require 'spec_helper'
 
-describe 'rvm::rubies' do
+describe 'rvm_sl::rubies' do
   let(:home) { '/home/vagrant' }
 
   let(:chef_run) do
     ChefSpec::ServerRunner.new(platform: 'ubuntu', version: '14.04') do |node|
-      node.set['rvm']['user']['name'] = 'vagrant'
-      node.set['rvm']['user']['password'] = 'vagrant'
+      node.set['rvm_sl']['user']['name'] = 'vagrant'
+      node.set['rvm_sl']['user']['password'] = 'vagrant'
     end.converge described_recipe
   end
 
@@ -14,8 +14,8 @@ describe 'rvm::rubies' do
     expect(chef_run).to run_execute('install_ruby')
   end
 
-  it 'runs a execute to modify permissions on rvm install dir' do
-    expect(chef_run).to run_execute('chown_rvm_dir')
+  it 'runs a execute to modify permissions on rvm_sl install dir' do
+    expect(chef_run).to run_execute('chown_rvm_sl_dir')
   end
 
   it 'converges successfully' do