+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-rvm CHANGELOG
+rvm_sl CHANGELOG
 =============
 
-This file is used to list changes made in each version of the rvm cookbook.
+This file is used to list changes made in each version of the rvm_sl cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of rvm
+- [your_name] - Initial release of rvm_sl
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.